infrastructure/requirements.txt
aws-cdk-lib==2.210.0
constructs>=10.0.0
boto3>=1.26.0
python-dotenv>=1.0.0
